pyproject.toml
[tool.poetry]
name = "websocket-p21"
version = "0.1.0"
description = ""
authors = ["xolmomin <xolmomin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
channels = "^4.1.0"
daphne = "^4.1.2"
channels-redis = "^4.2.0"
ujson = "^5.10.0"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
